requirements.txt
langchain
openai>=1.0.0
tinydb
pillow
python-dotenv
pyrogram
langchain-openai
